requirements.txt
pandas==1.5.3
streamlit==1.18.1
scikit-learn==1.2.1
tensorflow==2.12.0rc1
numpy==1.23.5
seaborn==0.12.2
Pillow==9.4.0